separate document from footnote schema
