+++ document/static/js/es6_modules/editor/schema.js
@@ -151,22 +151,9 @@ Footnote.register("command", "insert", {
     }
 })
 
-class FootnoteContainer extends Block {
-}
 
-FootnoteContainer.register("parseDOM", "div", {
-    parse: function(dom, state) {
-        if (!dom.classList.contains('footnote-container')) return false
-        state.wrapIn(dom, this)
-    }
-})
 
-FootnoteContainer.prototype.serializeDOM = (node, serializer) =>
-    serializer.renderAs(node, "div", {
-        class: 'footnote-container'
-    })
-
-class Citation extends Inline {
+export class Citation extends Inline {
     get attrs() {
         return {
             bibFormat: new Attribute({
@@ -244,7 +231,7 @@ Citation.register("command", "insert", {
     }
 })
 
-class Equation extends Inline {
+export class Equation extends Inline {
     get attrs() {
         return {
             equation: new Attribute({
@@ -294,7 +281,7 @@ Equation.register("command", "insert", {
 })
 
 
-class Figure extends Block {
+export class Figure extends Block {
     get attrs() {
         return {
             equation: new Attribute({
@@ -458,7 +445,6 @@ CommentMark.prototype.serializeDOM = (mark, serializer) => {
 
 export const fidusSchema = new Schema({
   nodes: {
-    //doc: {type: Doc, content: "part+"},
     doc: {type: Doc, content: "title subtitle authors abstract keywords body"},
     title: {type: Title, content: "text*", group: "part"},
     subtitle: {type: Subtitle, content: "text*", group: "part"},
@@ -492,37 +478,5 @@ export const fidusSchema = new Schema({
     link: LinkMark,
     code: CodeMark,
     comment: CommentMark
-  }
-})
-
-export const fidusFnSchema = new Schema({
-  nodes: {
-    doc: {type: Doc, content: "part+"},
-
-    footnote_end: {type: HorizontalRule, group: "part"},
-    footnotecontainer: {type: FootnoteContainer, content: "block+", group: "part"},
-
-    paragraph: {type: Paragraph, content: "inline<_>*", group: "block"},
-    heading: {type: Heading, content: "inline<_>*", group: "block"},
-    code_block: {type: CodeBlock, content: "text*", group: "block"},
-
-    blockquote: {type: BlockQuote, content: "block+", group: "block"},
-    ordered_list: {type: OrderedList, content: "list_item+", group: "block"},
-    bullet_list: {type: BulletList, content: "list_item+", group: "block"},
-    list_item: {type: ListItem, content: "block+", group: "block"},
-    horizontal_rule: {type: HorizontalRule, group: "block"},
-    figure: {type: Figure, group: "block"},
-
-    text: {type: Text, group: "inline"},
-    hard_break: {type: HardBreak, group: "inline"},
-    citation: {type: Citation, group: "inline"},
-    equation: {type: Equation, group: "inline"}
-  },
-
-  marks: {
-    em: EmMark,
-    strong: StrongMark,
-    link: LinkMark,
-    code: CodeMark
   }
 })
